modify: modify the folder structure
