+++ src/Components/Tasks/AddOrEditTask/AddEditTask.tsx
@@ -1,5 +1,5 @@
-import Navbar from "../../Components/Navbar/Navbar";
-import AddEditForm from "../../Components/AddOrEditTask/AddEditForm";
+import Navbar from "../Navbar/Navbar";
+import AddEditForm from "./AddEditForm";
 import { useParams } from "react-router-dom";
 
 export default function AddEditTask() {
