+++ src/Components/Tasks/context/TaskContext/TaskProvider.tsx
@@ -1,12 +1,21 @@
 import { useState } from "react";
 import { Task } from "../../types/TaskType/types";
-import { TaskContext } from "./TaskContext";
+import { createContext, useContext } from "react";
+
+export type TasksArrayType = {
+  tasks: Task[];
+  setTasks: (c: Task[]) => void;
+};
+const TaskContext = createContext<TasksArrayType>({
+  tasks: [],
+  setTasks: () => {},
+});
 
 type ChildrenType = {
   children: React.ReactNode;
 };
 
-export default function TaskProvider({ children }: ChildrenType) {
+export function TaskProvider({ children }: ChildrenType) {
   if (!localStorage.getItem("tasks-array")) {
     localStorage.setItem("tasks-array", JSON.stringify([]));
   }
@@ -20,4 +29,14 @@ export default function TaskProvider({ children }: ChildrenType) {
       </TaskContext.Provider>
     </div>
   );
+}
+
+export default function useTask() {
+  const context = useContext(TaskContext);
+
+  if (!context) {
+    throw new Error("useTheme must be used within a ThemeProvider");
+  }
+
+  return context;
 }
